Split driver production paths by state machine
